track subscription and improve testing
